Refactor to separate users and countries
This makes it possible for a user to manage multiple countries, and also
separates the login process from the country creation process.

This change allows us to use Twitter as an auth mechanism so that the
user can login with Twitter and then manage their countries and any
pending requests for them.