Make it easier to scroll the people list
In browsers that support CSS calc(), the people list now
fills the entire space to the right of the timeline,
which means you can more easily scroll it. Fixes #40.

In browsers that don’t support calc(), the people list
will be 15em wide, and pinned to the right edge of the
window. The timeline will be 30em, and there’ll be a white
gap between them.